requirements/requirements.txt
discord.py==2.1.0
Pillow==9.1.0
python-dotenv==0.21.0
